Overall Improvements
Changed UART baud rate from 9600 to 115200 and split some more of the code into functions and called what is needed. UART functions will be changed next, making them all generic functions that will be given relevent data from other functions. For instance, I am going to change "HEADER" to "UART_PRINTSTR" that will just print a whole string of data and then a "HEADER" function will load the data needed for the PRINTSTR function to work. Still rev 0.1.0 because it is not a working shell yet.